requirements.txt
PyNaCl>=1.5.0
pytest>=7.0.0
cryptography>=41.0.0